[SPEC] Update testing configuration
The test configurations have been updated to allow for tests to be run
in the Docker container. Also updated the CI with an alternate
installation of Elixir and depenencies to get the database to properly
build and run tests in the Github workflow.

README was also updated to reflect changes in testing execution