[FEAT] Update Destination CRUD Views
Biggest concern was to get the `/index` view into a manageable state. To
do this, I had to find ways to consolidate the related boolean fields,
like `travel time` and `season`.

The user has an option to go directly to an `/index` sorted by the
season, so that takes care of 4 columns. I eliminated the GPS
coordinates, as they shouldn't be necessary in an index view. I also
created several helpers to help combine other columns. The `container`
width was also increased from 90rem to 120rem.