[FEAT] Add Google Maps link to `destinations/show`
This is a solution that might be updated in the future. But for now
the most efficient way to get directions to a spot is to send the user
to Google Maps. It is also beneficial because a good majority of people
use Google Maps and can then utlize their history and saved places.