[FEAT] Add map to destination index page
Created `multiDestinationMap.js` to handle map with multiple
destinations and added it to `webpack`.

Created new actions in `mapping.ex` to gather data necessary for map.
Created two functions for the median coordinate. The first handles when
no destinations are present, returning default values, the second
handling when destinations are present, and returning the average
coordinate value.